Replaced ASMEventHandler with an equivalent based on lambdas
